pyproject.toml
[project]
name = "the-automaton-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langgraph>=1.0.9",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pypdf>=6.7.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "typing-extensions>=4.15.0",
]
